Initial DTLS implementation (plaintext broken).

+++ test/ofp/protocolrange_unittest.cpp
@@ -13,14 +13,14 @@ TEST(protocolrange, valid_empty) {
   EXPECT_TRUE(IsProtocolRangeValid(0, r1, 2, 8, &context1));
 
   UInt64 d2 = 0;
-  ByteRange r2{&d2, (size_t)0};
+  ByteRange r2{&d2, 0UL};
   Validation context2{nullptr};
   EXPECT_TRUE(IsProtocolRangeValid(0, r2, 2, 8, &context2));
 }
 
 TEST(protocolrange, invalid_empty) {
   UInt8 d1[2];
-  ByteRange r1{&d1[1], (size_t)0};
+  ByteRange r1{&d1[1], 0UL};
   Validation context1{nullptr};
   EXPECT_FALSE(IsProtocolRangeValid(0, r1, 2, 8, &context1));
 }
